+++ clients/admin-ui/src/features/datastore-connections/system_portal_config/DeleteConnectionModal.tsx
@@ -19,20 +19,18 @@ import React from "react";
 import { TrashCanSolidIcon } from "~/features/common/Icon/TrashCanSolidIcon";
 
 type DataConnectionProps = {
-  connectionKey: string;
-  onDelete: (connection_key: string) => void;
+  onDelete: () => void;
   deleteResult: any;
 };
 
 const DeleteConnectionModal: React.FC<DataConnectionProps> = ({
-  connectionKey,
   onDelete,
   deleteResult,
 }) => {
   const { isOpen, onOpen, onClose } = useDisclosure();
 
   const handleDeleteConnection = () => {
-    onDelete(connectionKey);
+    onDelete();
   };
 
   const closeIfComplete = () => {