+++ clients/admin-ui/src/features/system/system.slice.ts
@@ -126,6 +126,13 @@ const systemApi = baseApi.injectEndpoints({
       }),
       providesTags: ["Datamap", "System", "Datastore Connection"],
     }),
+    deleteSystemConnectionConfig: build.mutation({
+      query: (systemFidesKey) => ({
+        url: `/system/${systemFidesKey}/connection`,
+        method: "DELETE",
+      }),
+      invalidatesTags: () => ["Datastore Connection", "System"],
+    }),
   }),
 });
 
@@ -137,6 +144,7 @@ export const {
   useDeleteSystemMutation,
   useUpsertSystemsMutation,
   usePatchSystemConnectionConfigsMutation,
+  useDeleteSystemConnectionConfigMutation,
   useGetSystemConnectionConfigsQuery,
   usePatchSystemConnectionSecretsMutation,
 } = systemApi;